[fix](Nereids) move tables from connect context to statement context
